feat: evaluator module - Move segment rule evaluators to flag_engine.segments.evaluator module - Add missing runtime type checks
