+++ flag_engine/segments/evaluator.py
@@ -1,12 +1,24 @@
+import operator
+import re
 import typing
+from contextlib import suppress
 
+import semver
+
+from flag_engine.environments.models import EnvironmentModel
 from flag_engine.identities.models import IdentityModel
+from flag_engine.identities.traits.models import TraitModel
+from flag_engine.identities.traits.types import TraitValue
+from flag_engine.segments import constants
+from flag_engine.segments.models import (
+    SegmentConditionModel,
+    SegmentModel,
+    SegmentRuleModel,
+)
+from flag_engine.segments.types import ConditionOperator
 from flag_engine.utils.hashing import get_hashed_percentage_for_object_ids
-
-from ..environments.models import EnvironmentModel
-from ..identities.traits.models import TraitModel
-from . import constants
-from .models import SegmentConditionModel, SegmentModel, SegmentRuleModel
+from flag_engine.utils.semver import is_semver
+from flag_engine.utils.types import get_casting_function
 
 
 def get_identity_segments(
@@ -79,6 +91,7 @@ def _traits_match_segment_condition(
     identity_id: typing.Union[int, str],
 ) -> bool:
     if condition.operator == constants.PERCENTAGE_SPLIT:
+        assert condition.value
         float_value = float(condition.value)
         return (
             get_hashed_percentage_for_object_ids([segment_id, identity_id])
@@ -95,4 +108,102 @@ def _traits_match_segment_condition(
     if condition.operator == constants.IS_SET:
         return trait is not None
 
-    return condition.matches_trait_value(trait.trait_value) if trait else False
+    return _matches_trait_value(condition, trait.trait_value) if trait else False
+
+
+def _matches_trait_value(
+    condition: SegmentConditionModel,
+    trait_value: TraitValue,
+) -> bool:
+    if match_func := MATCH_FUNCS_BY_OPERATOR.get(condition.operator):
+        return match_func(condition.value, trait_value)
+
+    maybe_semver_trait_value: typing.Union[
+        TraitValue,
+        semver.VersionInfo,
+    ] = trait_value
+    if operator_func := OPERATOR_FUNCS_BY_OPERATOR.get(condition.operator):
+        with suppress(TypeError, ValueError):
+            if isinstance(maybe_semver_trait_value, str) and is_semver(condition.value):
+                maybe_semver_trait_value = semver.VersionInfo.parse(
+                    maybe_semver_trait_value,
+                )
+            match_value = get_casting_function(maybe_semver_trait_value)(
+                condition.value,
+            )
+            return operator_func(
+                maybe_semver_trait_value,
+                match_value,
+            )
+
+    return False
+
+
+def _evaluate_not_contains(
+    segment_value: typing.Optional[str],
+    trait_value: TraitValue,
+) -> bool:
+    return isinstance(trait_value, str) and str(segment_value) not in trait_value
+
+
+def _evaluate_regex(
+    segment_value: typing.Optional[str],
+    trait_value: TraitValue,
+) -> bool:
+    return (
+        trait_value is not None
+        and re.compile(str(segment_value)).match(str(trait_value)) is not None
+    )
+
+
+def _evaluate_modulo(
+    segment_value: typing.Optional[str],
+    trait_value: TraitValue,
+) -> bool:
+    if not isinstance(trait_value, (int, float)):
+        return False
+
+    if segment_value is None:
+        return False
+
+    try:
+        divisor_part, remainder_part = segment_value.split("|")
+        divisor = float(divisor_part)
+        remainder = float(remainder_part)
+    except ValueError:
+        return False
+
+    return trait_value % divisor == remainder
+
+
+def _evaluate_in(segment_value: typing.Optional[str], trait_value: TraitValue) -> bool:
+    if segment_value:
+        if isinstance(trait_value, str):
+            return trait_value in segment_value.split(",")
+        elif isinstance(trait_value, int) and not any(
+            trait_value is x for x in (False, True)
+        ):
+            return str(trait_value) in segment_value.split(",")
+    return False
+
+
+MATCH_FUNCS_BY_OPERATOR: typing.Dict[
+    ConditionOperator, typing.Callable[[typing.Optional[str], TraitValue], bool]
+] = {
+    constants.NOT_CONTAINS: _evaluate_not_contains,
+    constants.REGEX: _evaluate_regex,
+    constants.MODULO: _evaluate_modulo,
+    constants.IN: _evaluate_in,
+}
+
+OPERATOR_FUNCS_BY_OPERATOR: typing.Dict[
+    ConditionOperator, typing.Callable[..., bool]
+] = {
+    constants.EQUAL: operator.eq,
+    constants.GREATER_THAN: operator.gt,
+    constants.GREATER_THAN_INCLUSIVE: operator.ge,
+    constants.LESS_THAN: operator.lt,
+    constants.LESS_THAN_INCLUSIVE: operator.le,
+    constants.NOT_EQUAL: operator.ne,
+    constants.CONTAINS: operator.contains,
+}
